==> 2_wflow_sbm/03_create_calibration_config_files_wflow.ipynb <==
# Generate calibration forcing files

from glob import glob
from pathlib import Path

import toml

## Set Paths

# Snellius paths
ROOT = Path('/gpfs/work1/0/wtrcycle/users/jaerts/camels_uk/')
MODELS = Path(f'{ROOT}/wflow/data/')

## Set config

# Get available basin IDs wflow_sbm
basin_dirs = glob(f'{MODELS}/*')
basin_ids = [s.split('/')[-1] for s in basin_dirs]
basin_ids.sort()

# Set calibration interval
calibration_values = [1,5,10,15,20,
                      25,30,35,40,45,
                      50,55,60,65,70,
                      75,80,85,90,95,
                      100,125,150,175,
                      200,225,250,275,
                      300,350,400,450,550,
                      600,650,700,750,800,
                      850,900,950,1000,1500,
                      2000,2500,3000,4000,
                      4500,5000,7500,10000]

# Generate config_files

# Calibration Period
for basin_id in basin_ids:
    for calibration_value in calibration_values:
        config_file = toml.load(f'{MODELS}/{basin_id}/wflow_sbm.toml')
        config_file['loglevel'] = "debug"      
        config_file['path_log'] = f"log_ksathorfrac_{calibration_value}.txt"    
        config_file['starttime'] = '2000-01-01T00:00:00'
        config_file['endtime'] = '2008-12-31T00:00:00'

        config_file['input']['path_forcing'] = f'ceh-gear_chess_camels-gb_{basin_id}_2000_2017.nc'
        config_file['input']['path_static'] = f'staticmaps_ksathorfrac_{calibration_value}.nc'
        
        config_file['state']['path_output'] = f'ksathorfrac_{calibration_value}/outstate/outstates.nc'
        config_file['output']['path'] = f'ksathorfrac_{calibration_value}/output.nc'
        config_file['csv']['path'] = f'ksathorfrac_{calibration_value}/output.csv'
        
        with open(f'{MODELS}/{basin_id}/wflow_sbm_ksathorfrac_{calibration_value}.toml', 'w') as f:
            toml.dump(config_file, f)
